TorquePID .create(int) -> .create(double)

+++ control/TorquePID.java
@@ -36,7 +36,7 @@ public final class TorquePID {
         hasIntegralZone = b.hasIntegralZone;
     }
 
-    public static final Builder create(final int p) {
+    public static final Builder create(final double p) {
         return new Builder(p);
     }
